llvmPackages_{18,19}.mlir: fix build with gcc15
- add patch from 2 upstream commits that were not backported to llvm_18
and llvm_19:

https://github.com/llvm/llvm-project/commit/41eb186fbb024898bacc2577fa3b88db0510ba1f
Fixes errors with gcc15:
/build/mlir-src-19.1.7/mlir/include/mlir/Dialect/Affine/IR/ValueBoundsOpInterfaceImpl.h:31:11:
error: 'int64_t' was not declared in this scope

https://github.com/llvm/llvm-project/commit/101109fc5460d5bb9bb597c6ec77f998093a6687
Fixes errors with gcc15:
/build/mlir-src-18.1.8/mlir/include/mlir/Target/SPIRV/Deserialization.h:29:51:
error: 'uint32_t' was not declared in this scope

Cannot use `fetchpatch` for first because patch does not apply on llvm_18
or llvm_19. Rebased patch on either version also does no apply to the
other.
So use 2 versions of combined patches with 2 commits.